Move TRI and VRM functions from helper to coupling.terrain
