+++ mmctools/coupling/terrain.py
@@ -506,3 +506,119 @@ def calc_slope(x,y,z):
     rise_run = np.sqrt(dz_dx**2 + dz_dy**2)
     slope[1:-1,1:-1] = np.degrees(np.arctan(rise_run))
     return slope
+
+
+def calcTRI(hgt,window=None,footprint=None):
+    '''
+    Terrain Ruggedness Index
+    Riley, S. J., DeGloria, S. D., & Elliot, R. (1999). Index that 
+        quantifies topographic heterogeneity. intermountain Journal 
+        of sciences, 5(1-4), 23-27.
+    
+    hgt : array
+        Array of heights over which TRI will be calculated
+    window : int
+        Length of window in x and y direction. Must be odd.
+    '''
+    from scipy.ndimage.filters import generic_filter
+
+    # Window setup:
+    if footprint is not None:
+        assert window is None, 'Must specify either window or footprint'
+        window = np.shape(footprint)[0]
+    
+    assert (window/2.0) - np.floor(window/2.0) != 0.0, 'window must be odd...'
+    Hwindow = int(np.floor(window/2))
+    
+    # Type and dimension check:
+    if isinstance(hgt,(xr.Dataset,xr.DataArray,xr.Variable)):
+        hgt = hgt.data    
+    assert len(np.shape(hgt)) == 2, 'hgt must be 2-dimensional. Currently has {} dimensions'.format(len(np.shape(hgt)))
+    
+    ny,nx = np.shape(hgt)
+    
+    def tri_filt(x):
+        middle_ind = int(len(x)/2)
+        return((sum((x - x[middle_ind])**2.0))**0.5)
+    
+    if footprint is None:
+        tri = generic_filter(hgt,tri_filt, size = (window,window))
+    else:
+        tri = generic_filter(hgt,tri_filt, footprint=footprint)
+    
+    return tri
+
+
+def calcVRM(hgt,res,window=None,footprint=None,fill_depressions=True,return_slope_aspect=False):
+    '''
+    Vector Ruggedness Measure
+    Sappington, J. M., Longshore, K. M., & Thompson, D. B. (2007). 
+        Quantifying landscape ruggedness for animal habitat analysis: 
+        a case study using bighorn sheep in the Mojave Desert. The 
+        Journal of wildlife management, 71(5), 1419-1426.
+    
+    hgt : array
+        Array of heights over which TRI will be calculated
+    res : int or float
+        Resolution of the underlying hgt array. Should be constant in x and y
+        Needed for proper slope calculation
+    window : int
+        Length of window in x and y direction. Must be odd.
+    '''
+    import richdem as rd
+    from scipy.ndimage.filters import generic_filter
+
+    # Window setup:
+    if footprint is not None:
+        assert window is None, 'Must specify either window or footprint'
+        window = np.shape(footprint)[0]
+        
+    assert (window/2.0) - np.floor(window/2.0) != 0.0, 'window must be odd...'
+    Hwndw = int(np.floor(window/2))
+
+    # Type and dimension check:
+    if isinstance(hgt,(xr.Dataset,xr.DataArray,xr.Variable)):
+        hgt = hgt.data    
+    assert len(np.shape(hgt)) == 2, 'hgt must be 2-dimensional. Currently has {} dimensions'.format(len(np.shape(hgt)))
+    ny,nx = np.shape(hgt)
+
+    # Determine scale based on resolution of hgt array
+    zscale = 1/res
+
+    # Get slope and aspect:
+    hgt_rd = rd.rdarray(hgt, no_data=-9999)
+    if fill_depressions:
+        rd.FillDepressions(hgt_rd, in_place=True)
+    slope  = rd.TerrainAttribute(hgt_rd, attrib='slope_degrees',zscale=zscale)
+    aspect = rd.TerrainAttribute(hgt_rd, attrib='aspect')
+    
+    # Calculate vectors:
+    vrm = np.zeros((ny,nx))
+    rugz   = np.cos(np.deg2rad(slope))
+    rugdxy = np.sin(np.deg2rad(slope))
+    rugx   = rugdxy*np.cos(np.deg2rad(aspect))
+    rugy   = rugdxy*np.sin(np.deg2rad(aspect))
+
+    def vrm_filt(x):
+        return(sum(x)**2)
+
+    if footprint is None:
+        vrmX = generic_filter(rugx,vrm_filt, size = (window,window))
+        vrmY = generic_filter(rugy,vrm_filt, size = (window,window))
+        vrmZ = generic_filter(rugz,vrm_filt, size = (window,window))
+    else:
+        vrmX = generic_filter(rugx,vrm_filt, footprint=footprint)
+        vrmY = generic_filter(rugy,vrm_filt, footprint=footprint)
+        vrmZ = generic_filter(rugz,vrm_filt, footprint=footprint)    
+
+
+    if footprint is not None:
+        num_points = len(footprint[footprint != 0.0])
+    else:
+        num_points = float(window**2)
+    vrm = 1.0 - np.sqrt(vrmX + vrmY + vrmZ)/num_points
+    if return_slope_aspect:
+        return vrm,slope,aspect
+    else:
+        return vrm
+
